+++ libp2p/src/bandwidth.rs
@@ -43,9 +43,11 @@ use std::{
 pub(crate) struct BandwidthLogging<SMInner> {
     #[pin]
     inner: SMInner,
+    #[allow(deprecated)]
     sinks: Arc<BandwidthSinks>,
 }
 
+#[allow(deprecated)]
 impl<SMInner> BandwidthLogging<SMInner> {
     /// Creates a new [`BandwidthLogging`] around the stream muxer.
     pub(crate) fn new(inner: SMInner, sinks: Arc<BandwidthSinks>) -> Self {
@@ -101,11 +103,13 @@ where
 }
 
 /// Allows obtaining the average bandwidth of the streams.
+#[deprecated(note = "Use `libp2p::SwarmBuilder::with_bandwidth_metrics` or `libp2p_metrics::BandwidthMetricTransport` instead.")]
 pub struct BandwidthSinks {
     inbound: AtomicU64,
     outbound: AtomicU64,
 }
 
+#[allow(deprecated)]
 impl BandwidthSinks {
     /// Returns a new [`BandwidthSinks`].
     pub(crate) fn new() -> Arc<Self> {
@@ -137,6 +141,7 @@ impl BandwidthSinks {
 pub(crate) struct InstrumentedStream<SMInner> {
     #[pin]
     inner: SMInner,
+    #[allow(deprecated)]
     sinks: Arc<BandwidthSinks>,
 }
 
@@ -148,6 +153,7 @@ impl<SMInner: AsyncRead> AsyncRead for InstrumentedStream<SMInner> {
     ) -> Poll<io::Result<usize>> {
         let this = self.project();
         let num_bytes = ready!(this.inner.poll_read(cx, buf))?;
+        #[allow(deprecated)]
         this.sinks.inbound.fetch_add(
             u64::try_from(num_bytes).unwrap_or(u64::max_value()),
             Ordering::Relaxed,
@@ -162,6 +168,7 @@ impl<SMInner: AsyncRead> AsyncRead for InstrumentedStream<SMInner> {
     ) -> Poll<io::Result<usize>> {
         let this = self.project();
         let num_bytes = ready!(this.inner.poll_read_vectored(cx, bufs))?;
+        #[allow(deprecated)]
         this.sinks.inbound.fetch_add(
             u64::try_from(num_bytes).unwrap_or(u64::max_value()),
             Ordering::Relaxed,
@@ -178,6 +185,7 @@ impl<SMInner: AsyncWrite> AsyncWrite for InstrumentedStream<SMInner> {
     ) -> Poll<io::Result<usize>> {
         let this = self.project();
         let num_bytes = ready!(this.inner.poll_write(cx, buf))?;
+        #[allow(deprecated)]
         this.sinks.outbound.fetch_add(
             u64::try_from(num_bytes).unwrap_or(u64::max_value()),
             Ordering::Relaxed,
@@ -192,6 +200,7 @@ impl<SMInner: AsyncWrite> AsyncWrite for InstrumentedStream<SMInner> {
     ) -> Poll<io::Result<usize>> {
         let this = self.project();
         let num_bytes = ready!(this.inner.poll_write_vectored(cx, bufs))?;
+        #[allow(deprecated)]
         this.sinks.outbound.fetch_add(
             u64::try_from(num_bytes).unwrap_or(u64::max_value()),
             Ordering::Relaxed,