+++ misc/metrics/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.14.0 - unreleased
 
+- Add `_ => {}` arm to `#[non_exhaustive]` enum `match`es
+  See [PR 4581](https://github.com/libp2p/rust-libp2p/pull/4581).
 - Add metrics for `SwarmEvent::{NewExternalAddrCandidate,ExternalAddrConfirmed,ExternalAddrExpired}`.
   See [PR 4721](https://github.com/libp2p/rust-libp2p/pull/4721).